Handle bveCVX in ibBTC peak (non-gnosis contract)
Added the ibBTC peak contract where the treasury accidentally sent some bveCVX a while back.  We have been voting with it/it belongs to the treasury, and we are waiting to upgrade the contract to sweep it back.  These funds are actually the property of the treasury-voter multisig and vote with the rest of our bveCVX, so we have attributed the balance of this non-gnosis contract back to our treasury-voter multisig where we can claim/lock and vote from.